enforece holy import trinity using goimports
This PR changes our gofmt check to
goimports -local github.com/sapcc/kubernikus.

This caused the tests to fail if imports are not ordered in the following way:
```
stdlib
[empty line]
external deps
[empty line]
kubernikus internal deps
````

I purposly excluded the generated files from this check to avoid confusion when we regenerate stuff.