Add leveled logging (#129)
* Add leveled logging

I reused the `—v` flag for glog to set the desiered log level. By default we only log messages with no level or level 0.
For the flag processing to work I had to move the logger creation into the cobra commands `Run` functions because otherwise I can’t access the flag without hussle.

I moved the logging setup to logutil.NewLogger() to that we can reuse that function in all binaries.

* Remove unused functions

* Refactor wormhole server

use one flat Options struct for both the controller and the tunnel.
They are not meant to be run independently and share some things (e.g. logger)

Also consolidated the server in a single package because