Upgrade to Kubernetes 1.8.5 (#132)
* Upgrade to Kubernetes 1.8.5

This commit upgrades the helm chart and ignition template to use 1.8.5.

Currently, this assumes that klusters without version information in the
spec are on 1.7.5. All new klusters must provide a version.

With 1.8.5 the default behaviour of the Openstack LoadBalancer creation
changed. Instead of defaulting to an internal LoadBalancer it now
creates an external LB by default. This requires the external network to
be configured in the cloud-provider config.

Fixes #62

* removes v from versions

* fixes logging in case of error while creating kluster struct

* uses one version for new klusters only