Sort (and filter) flavors in metadata response
Flavors are now sorted by RAM and VCPUs.
Also the RAM and VCPUs are not part of the reponse.
Additionally I hardcoded the flavors list call for now to only return flavors with 2000 MB RAM or more to get rid of the tiny images that don’t work with our default coreos image

Fixes #130