Use Finalizers for Orchestrating Kluster Deletion
This refactors the orchestration for Kluster Deletion. It makes use of
the Finalizers concept. It allows to shift the responsibility of
cleaning up to individual controllers. It spares us the implementation
of a more complicated kluster-phase state machine.

See: https://kubernetes.io/docs/tasks/access-kubernetes-api/extend-api-custom-resource-definitions/#finalizers

We need this as a prerequisite for the implementation of the Deorbiter
in #144.