Restrict cluster name to maximum length 20
The maximum length comes from limitations in kubernetes (max 63 chars for resource names).
We concatinate with the projectid (32 chars): `$name-$projectid` which results in a maximum of 53 chars for the qualified name (e.g. helm release). This leaves us with 10 chars for suffixes for resource in our kube-master helm chart.