Add Controller for Deorbiting Klusters
The Kubernetes `controller-manager` runs controllers that are responsible for
creating, updating and deleting of OpenStack resources. It auto-provisions
Cinder volumes for Persistent Volume Claims, load balancers for services and
routes for nodes in the cluster.  When a kluster is terminated it is desired
that these resources are cleaned up automatically.

The problem is that the `controller-manager` only cleans up, when the
respective resources are deleted from the cluster.

This is where the `deorbiter` comes into play. It inspects the cluster and
deletes all PVCs that are backed by Cinder, as well as all services of type
`LoadBalancer`. Afterwards, it waits until the `controller-manager` has
completed the clean-up.

Fixes #45.