Use nodeobservatory to update node pool status
This change refactors the NodeObservatory to act more like an informer instead of a controller. It also adds a factory for it so that the NodeObservatory only started when it actually request by a controller.
Secondly launchctl now uses the NodeObservatory to update the `Healthy` and `Schedulable` status of node pools.