Probe ingresses in kubernikus-system over region_probed_from
We probe all ingresses in kubernikus-system every five seconds with the
blackbox-exporter/prober from all the regions mentioned in the list
region_probed_from.
If the region the ingress is running in is part of the list it will
be treated specially and probe from inside the cluster using the
internal service instead of the ingress.

This was done to not be flooded in redundant probes while stile having
a probe from outside of the region whose number can be raised easily.
The rationale behind the handpicking of the region_probed_from elements
is to be able to define a region that is nearby.