Adds quota checks pre/post e2e test runs
This commit adds a quota check before the e2e test are started. The test
fail with a concrete error message which quota was exceeded. That
safe-guards against obscure timeouts that happen to to missing quota.

After the e2e test have been completed, there's now also an additional
post-flight quota check. It will fail the test if any resource is still
left over and hasn't been cleaned up properly.

This should make it easier to identify when, where and how the tests are
leaking resources.

In case an existing kluster is reused for e2e tests, the quota checks
are skipped.