Adds Servicing Controller
This controller handles node upgrades when the Kubernetes or CoreOS versions
are changed. It gracefully drains nodes before performing any action.

For Kubernetes upgrades the strategy is to replace the fleet by terminating
the nodes. CoreOS updates are handled by a soft reboot.

In order to allow the payload to settle only a single node per cluster is
processed at a time. Inbetween updates there's a 1h grace period.

In case any node in the cluster is unhealthy the upgrades are skipped. This
is to safeguard against failed upgrades destroying the universe.

For rollout and testing purposed the node upgrades are disabled by default.
They can manually be enabled by setting the node annotaion:

   kubernikus.cloud.sap/servicing=true