fix cache bug in flightcontroller
This fixes an issue where our KlusterClient cached a specific version of the kluster object. This caused updates to the security group not to be reflected in the flight controller until the operator was restarted.

With this change the kluster object is not stored in the KlusterClient anymore but needs to be passed to the client functions.