Adds suppository migration style. Fixes Root Certificate (#411)
This PR adds a utility function that helps to apply a script to all nodes in a cluster.

It creates a DaemonSet in a randomly suffixed namespace. It includes an init container that applies the provided script with a bit of `chroot` magic. It runs in `priviledged` mode and has access to binaries on the host. The host's filesystem is mapped into the init container to `/host` and used as the root for the `chroot` command. `chroot` executes the provided script, that is inserted into the pod as ConfigMap. If the init container successfully completes the `pause` container will wait endlessly for self-destruction.

Afterwards the function checks whether created Pods are `Running` indicating that the migration applied to all nodes. If so, everything is deleted. 

This can be used in a migration to apply something to all nodes. See the provided migration that fixes a broken certificate.