flag is deprecated and defaults to true now
This enables a smooth transition to PSP. Today, users would have to
manually set --allow-privileged to true before transitioning to PSP,
which isn't a smooth deprecation path for the flag (we want people
to stop setting it). This PR makes the default behavior isomorphic
with what will happen after the flag is removed.

https://github.com/kubernetes/kubernetes/pull/63442