update k8sniff deployment
This updates to k8sniff to our own build. Rebuilding with go 1.12.
This has a performance benefit as it uses the splice syscall to proxy.
And it also fixes hanging connections to backend because go 1.12 enables tcp keepalives by default.

I'm also updateing the deployment to have 2 replicas running with anti affinity to improve availability of cluster apis.