credential api: embed cluster version in cluster name
I tried using the extensions fields that are defined in the Config api type but couldn't get them to work. As soon as I add an extension kubectl can't merge kubeconfigs anymore. Asked in #sig-api-machinry and got no helpful response.

So this added the version to the cluster name name@version. Should be totaly invisible to the user and we can use this locally to switch to the correct kubectl version.

One caveat: As clusters get upgraded we will leave old cluster definitions dangeling in the kubeconfig but yolo.