turns Role into ClusterRole
There is no `Role` with the name `edit`. By default, users which are assigned the `kubernikus:member` RoleBinding, will get that the `edit` Role can't be found. 

This was either always broken. Or the `Role` was changed to a `ClusterRole` without us noticing. It is a user-facing role and comes with Kubernetes itself. For more info, see here: https://kubernetes.io/docs/reference/access-authn-authz/rbac/#user-facing-roles

Either way, it needs to be a `ClusterRole` now.