Adds Exceptions for FlatCar Replacements
A breaking change in FlatCar requires us to replace nodes even for an OS
update. This change allows nodes to be marked as "replacable" even if
under normal circumstances a reboot would be enough.

Specifically, all nodes with FlatCar < 2905 will be replaced if an OS or
Kubernetes upgrade is necessary.