Add support for custom root disks (#600)
This PR adds a `customRootDisk` field to the nodepool spec. If set the vms get an ephermeral cinder based root disk that is deleted when the vm is deleted.

This PR adds a `customRootDisk` field to the nodepool spec. If set the vms get an ephermeral cinder based root disk that is deleted when the vm is deleted.

Co-authored-by: Jan Knipper <9881823+jknipper@users.noreply.github.com>