Audit logging for API server (#629)
* Add flag to enable audit logging

* Add fluentd to ship audit logs

For now logs can be shipped to Elasticsearch. Endpoint, user and
password are passed via the Kubernikus API.

* Source ELK credentials from in-cluster secret

Credentials for elasticsearch cannot be passed via API.

* Add swift as backend for audit log fluentd

The fluentd-openstack plugin has two issues regarding authentification:
  1. Some auth parameters cannot be passed via config.
  2. fog-openstack the underlying openstack library does not pass
  openstack_project_name around correctly, so it is better to rely on
  IDs instead of names.
Both are adressed by patching the fluentd openstack plugin.

* Cleanup audit log values

* Update audit helm value when upgrading cluster

This requires openstack_project_domain_id and openstack_user_domain_id
to be present, which is ensured with a migration

* Improve fluentd Dockerfile

It's smaller and the swift api key is now hidden.

* Fix chart-tests

* Polish kube-mater chart

* Reconcile swift containers

The existence of backup and audit-log swift containers is checked during
the running phase of a kluster. These containers are then created if
required.

* Add fluentd images to v1.22.4 clusters

* Polish audit logging