Add more omitempty to openstackValues
The YAML marshaler changed with the Helm 2 to 3 migration. A struct,
which fields are all zero values will now produce a keys in a parent
struct even if the respective field in the parent struct is marked as
omitempty. By marking all fields in the nested struct as omitempty
"key: {}" will be generated, which should satisfy the .openstack empty
check in kube-master