Waits for Load Balancer Deletions via Finalizers (#542)
* waits for load balancer deletions via finalizers

This commit changes the wait behaviour when deleteing load balancers.
Previously service deleteion didn't use finalizers and no feedback was
given. The service disapeared immediately. The deorbiter used a fixed
timeout of 2min to wait for load balancer cleanup.

Since the loadbalancer driver in CCEE was changed to Octavia, that grace
period is not enough anymore. Deleteion takes longer and in case of 1.16
never completes if the LoadBalancer is not completely provisioned, e.g.
due to quota limitations. Especially a FIP is expected to be attached
and causes the deletion to never complete.

This fix depends on Kubernetes 1.17 to be fully functional. For
Kubernetes < 1.17 the Deorbiter will wait 24h for the service to be
deleted. For 1.16 it is possible to set the `use-octavia` flag as
a workaround.

* handle kluster < 1.17 gracefully

Co-authored-by: Fabian Ruff <fabian.ruff@sap.com>