Add support for structured authentication (#995)
* Initial support for structured-auth

Signed-off-by: Fabian Ruff <fabian.ruff@sap.com>

* add structured authenitcation configuration

This commit adds a new field “authenticationConfiguration” to the kluster spec that allowing to provide the content of the —authentication-configuration config file.
The ground controller reconciles any changes to the api and updates a configmap that is referenced in the apiserver deployment. As any changes to the configfile are automatically picked up by the apiserver this changes to the kluster spec become effective within a minute.

Signed-off-by: Fabian Ruff <fabian.ruff@sap.com>

* incorporate code-review feedback

Signed-off-by: Fabian Ruff <fabian.ruff@sap.com>

---------

Signed-off-by: Fabian Ruff <fabian.ruff@sap.com>