Fix CA regeneration
When regenerating the TLS CA certificate we must be careful to keep the subject of the new CA exactly the same byte for byte.

Otherwise the old CA is not considered in a cert pool when validating certificates issued by the new CA:
https://github.com/golang/go/blob/497cb7c0c3042d3c6605b46a1bf35b7c3bc8b046/src/crypto/x509/cert_pool.go#L144

How the subject is rended into bytes from a pkix.Name struct is not guaranteed to be stable across go versions.

We actually ran into this issue before and already filed a bug for this:
https://github.com/golang/go/issues/45882

Signed-off-by: Fabian Ruff <fabian.ruff@sap.com>